move the runner logic on another file
